Revert "try switching from diff-so-fancy to delta"
It caused too many colors, and it was no longer clear which lines were added or removed.
What previously was an all green file that had been added became a blue, yellow, and green file. It
wasn't even clear that it was syntax highlighting it. Maybe it's just not good for Ruby files? Or my
terminal colors aren't compatible?